+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - **nopt** has been upgraded from 4.0.3 to **5.0.0**
 - **semver** has been upgraded from 7.1.3 to **7.3.5**
 - **shortid** has been replaced by nanoid
+- **socket.io-client** has been upgraded from 2.3.0 to **4.1.2**
 
 # 9.0.2 / 2020-11-05